Move helper functions to separate shared file
This way they can be used by integration tests and unit tests
regardless of used tags
