+++ executors/kubernetes/helpers_kubernetes_test.go
@@ -1,0 +1,27 @@
+package kubernetes
+
+import (
+	restclient "k8s.io/client-go/rest"
+
+	"gitlab.com/gitlab-org/gitlab-runner/common"
+	"gitlab.com/gitlab-org/gitlab-runner/executors"
+)
+
+// GetKubeClientConfig is used to export the getKubeClientConfig function for integration tests
+func GetKubeClientConfig(config *common.KubernetesConfig) (kubeConfig *restclient.Config, err error) {
+	return getKubeClientConfig(config, new(overwrites))
+}
+
+// IsKubernetesPodNotFoundError is used to export the isKubernetesPodNotFoundError function for integration tests
+func IsKubernetesPodNotFoundError(err error) bool {
+	return isKubernetesPodNotFoundError(err)
+}
+
+// NewDefaultExecutorForTest is used to expose the executor to integration tests
+func NewDefaultExecutorForTest() common.Executor {
+	return &executor{
+		AbstractExecutor: executors.AbstractExecutor{
+			ExecutorOptions: executorOptions,
+		},
+	}
+}
